Edit seed

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,7 +22,7 @@ class CreateUsersTable extends Migration
             $table->rememberToken();
             $table->string('full_name')->nullable();
             $table->integer('age')->nullable();
-            $table->float('balance')->default(0);
+            $table->double('balance')->default(0);
             $table->string('avatar')->nullable();
             $table->text('certificate')->nullable();
             $table->integer('votes')->nullable();
